refactor: split store_mut.rs
